+++ conviva/src/main/java/com/bitmovin/analytics/conviva/ConvivaAnalyticsIntegration.java
@@ -86,35 +86,15 @@ public class ConvivaAnalyticsIntegration {
         this(player, customerKey, context, new ConvivaConfig());
     }
 
-    public ConvivaAnalyticsIntegration(Player player,
-                                       String customerKey,
-                                       Context context,
-                                       ConvivaConfig config) {
-        this(player, customerKey, context, config, null);
-    }
-
-    public ConvivaAnalyticsIntegration(Player player,
-                                       String customerKey,
-                                       Context context,
-                                       ConvivaConfig config,
-                                       ConvivaVideoAnalytics videoAnalytics
+    public ConvivaAnalyticsIntegration(
+        Player player,
+        String customerKey,
+        Context context,
+        ConvivaConfig config
     ) {
-        this(player, customerKey, context, config, videoAnalytics, null);
+        this(player, customerKey, context, config, null, null, null);
     }
 
-    public ConvivaAnalyticsIntegration(Player player,
-                                       String customerKey,
-                                       Context context,
-                                       ConvivaConfig config,
-                                       ConvivaVideoAnalytics videoAnalytics,
-                                       ConvivaAdAnalytics adAnalytics
-    ) {
-        this(player, customerKey, context, config, videoAnalytics, adAnalytics, null);
-    }
-
-    /**
-     * For testing purposes only.
-     */
     ConvivaAnalyticsIntegration(
         @Nullable Player player,
         String customerKey,